+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.2.2
+## Changed
+- Slashes in filenames are now ignored
 ## Fixed
+- Fix bug that when using braces in filenames
 - [malody] Dumping does not write placeholder `null` values anymore
 
 # v1.2.1